// src/types/NightViewSpot.d.ts
export interface NightViewSpot {
    TITLE: string;
    ADDR: string;
    LA: string;
    LO: string;
    TEL_NO: string;
    URL: string;
    OPERATING_TIME: string;
    ENTR_FEE: string;
    CONTENTS?: string;
    SUBWAY: string;
    BUS: string;
    MAIN_IMG?: string;
    SUBJECT_CD: string;
    NUM: string;
    PARKING_INFO?: string;
    REG_DATE?: string;
    MOD_DATE?: string;
    CONTENT: string;
}